refactor(reporters): move reporter selection bean to cli layer
Бин filteredReporters жил в reporters/infrastructure/ReportersConfiguration, но
зависел от cli.AnalyzeCommand (читал --reporter), создавая инверсию reporters→cli.
Отбор репортёров по ключам — это разбор CLI-ввода, поэтому бин перенесён в новый
cli/ReporterSelectionConfiguration. Имя бина (filteredReporters) и @Lazy сохранены,
поэтому @Qualifier-инъекция в ReportersAggregator и тайминг picocli не меняются.
Пустой пакет reporters/infrastructure удалён, reporters/CLAUDE.md обновлён.

Заморозка слоёв авто-сжалась: 5 → 3 (остался только diagnostics→providers через
IdenticalExpressionsDiagnostic→FormatProvider).
