+++ archunit_store/84350e2f-67b7-4a66-b776-0b0d67f02341
@@ -1,5 +1,3 @@
 Constructor <com.github._1c_syntax.bsl.languageserver.diagnostics.IdenticalExpressionsDiagnostic.<init>(com.github._1c_syntax.bsl.languageserver.providers.FormatProvider)> has parameter of type <com.github._1c_syntax.bsl.languageserver.providers.FormatProvider> in (IdenticalExpressionsDiagnostic.java:0)
 Field <com.github._1c_syntax.bsl.languageserver.diagnostics.IdenticalExpressionsDiagnostic.formatProvider> has type <com.github._1c_syntax.bsl.languageserver.providers.FormatProvider> in (IdenticalExpressionsDiagnostic.java:0)
 Method <com.github._1c_syntax.bsl.languageserver.diagnostics.IdenticalExpressionsDiagnostic.getOperandText(com.github._1c_syntax.bsl.languageserver.utils.expressiontree.BinaryOperationNode)> calls method <com.github._1c_syntax.bsl.languageserver.providers.FormatProvider.getNewText(java.util.List, java.util.Locale, org.eclipse.lsp4j.Range, int, org.eclipse.lsp4j.FormattingOptions)> in (IdenticalExpressionsDiagnostic.java:197)
-Method <com.github._1c_syntax.bsl.languageserver.reporters.infrastructure.ReportersConfiguration.filteredReporters(java.util.Collection, com.github._1c_syntax.bsl.languageserver.cli.AnalyzeCommand)> calls method <com.github._1c_syntax.bsl.languageserver.cli.AnalyzeCommand.getReportersOptions()> in (ReportersConfiguration.java:44)
-Method <com.github._1c_syntax.bsl.languageserver.reporters.infrastructure.ReportersConfiguration.filteredReporters(java.util.Collection, com.github._1c_syntax.bsl.languageserver.cli.AnalyzeCommand)> has parameter of type <com.github._1c_syntax.bsl.languageserver.cli.AnalyzeCommand> in (ReportersConfiguration.java:0)
